Refactor tests a little.
