+++ tests/conftest.py
@@ -28,12 +28,26 @@ import zipfile
 import numpy as np
 import pytest
 
+from delta.config import config
 from delta.imagery.sources import tiff
 
 import delta.config.modules
 delta.config.modules.register_all()
 
 assert 'tensorflow' not in sys.modules, 'For speed of command line tool, tensorflow should not be imported by config!'
+
+from delta.imagery import imagery_dataset #pylint: disable=wrong-import-position
+
+def config_reset():
+    """
+    Rests the configuration with useful default options for testing.
+    """
+    config.reset() # don't load any user files
+    config.load(yaml_str=
+                '''
+                mlflow:
+                  enabled: false
+                ''')
 
 def generate_tile(width=32, height=32, blocks=50):
     """Generate a widthXheightX3 image, with blocks pixels surrounded by ones and the rest zeros in band 0"""
@@ -105,3 +119,44 @@ NUM_SOURCES = 1
 @pytest.fixture(scope="session")
 def all_sources(worldview_filenames):
     return [(worldview_filenames, '.zip', 'worldview', '_label.tiff', 'tiff')]
+
+def load_dataset(source, output_size):
+    config_reset()
+    (image_path, label_path) = source[0]
+    config.load(yaml_str=
+                '''
+                io:
+                  cache:
+                    dir: %s
+                dataset:
+                  images:
+                    type: %s
+                    directory: %s
+                    extension: %s
+                    preprocess:
+                      enabled: false
+                  labels:
+                    type: %s
+                    directory: %s
+                    extension: %s
+                    preprocess:
+                      enabled: false
+                train:
+                  network:
+                    chunk_size: 3''' %
+                (os.path.dirname(image_path), source[2], os.path.dirname(image_path), source[1],
+                 source[4], os.path.dirname(label_path), source[3]))
+
+    dataset = imagery_dataset.ImageryDataset(config.dataset.images(), config.dataset.labels(),
+                                             config.train.network.chunk_size(), output_size,
+                                             config.train.spec().chunk_stride)
+    return dataset
+
+@pytest.fixture(scope="function", params=range(NUM_SOURCES))
+def dataset(all_sources, request):
+    source = all_sources[request.param]
+    return load_dataset(source, 1)
+
+@pytest.fixture(scope="function")
+def dataset_block_label(all_sources):
+    return load_dataset(all_sources[0], 3)
